- Point class's variable 'id' type was double, now it is integer. - Operators class's method 'findMinIndex' is fixed. - App class is updated.

+++ src/medianProblem/Point.java
@@ -1,6 +1,6 @@
 package medianProblem;
 public class Point{
-    public double id = 0;
+    public int id = 0;
     public double x = 0;
     public double y = 0;
     public double demand = 0;
